Respect `--project` in `uv version`
Previously, we were using the wrong `Workspace` discovery and would report the version of the workspace root, which would iterate up from the `--project` directory and return the workspace root (with or without a project in the root). Instead, we need `ProjectWorkspace` discovery that returns the closest project.

This fixes `uv version --project <path>` where `<path>` belongs to a workspace member.

Fixes #13213