Build backend: Allow escaping in globs
PEP 639 does not allow any characters that aren't in either their limited glob syntax or the alphanumeric Unicode characters. This means there's no way to express a glob such as `**/@test` for the excludes.

We extend the glob syntax from PEP 639 by introducing backslash escapes, which can escape all characters but path separators (forward and backwards slashes) to be parsed verbatim.

This means we have two glob parsers: The strict PEP 639 parser for `project.license-files`, and our extended parser for `tool.uv`, with a slight difference if you need to use special characters.