Apply first set of Rustfmt edition 2024 changes
Rustfmt introduces a lot of formatting changes in the 2024 edition. To not break everything all at once, we split out the set of formatting changes compatible with both the 2021 and 2024 edition by first formatting with the 2024 style, and then again with the currently used 2021 style.

Notable changes are the formatting of derive macro attributes and trailing semicolons in more places.