Clarify adding SSH Git dependencies
The current instructions say 

> prefix a Git-compatible URL (i.e., that you would use with git clone) with git+.

But this does not work with the URL that Github gives you when you choose Clone -> SSH via the UI, which is of the form `git@github.com:astral-sh/uv.git`. If you prefix this with `git+`, i.e.

`git+git@github.com:astral-sh/uv.git`

it does not work.