allow running non-default Python interpreters directly via uvx
Previously if you wanted to run e.g. PyPy via `uvx`, you had to spell it
like `uvx -p pypy python`. Now we share (some of) the
PythonRequest::parse machinery to handle the executable, so all of the
following examples work:

- uvx python
- uvx python3.8
- uvx 38
- uvx pypy
- uvx pypy@3.8
- uvx graalpy
- uvx graalpy@3.8

The `python` (and on Windows only, `pythonw`) special cases are
retained, which normally aren't allowed values of `-p`/`--python`.
However, the following examples deliberately *don't* work. Instead we
interpret these arguments as package names. (The `pp` package exists,
and we try to run it, but it fails to build.)

- uvx 38
- uvp pp
- uvp pp38

Closes https://github.com/astral-sh/uv/issues/13536.