Fix version json tests to work outside git checkout
Fix the two version json tests to account for the possibility that uv
was built outside a git checkout (e.g. from an unpacked git archive)
and therefore does not have the commit info available.  This approach
uses separate snapshots for the two cases, as suggested in discussion
of pull request #13251.

Fixes #13212