Make `requirements_txt_ssh_git_username` not write into homedir
Modify `requirements_txt_ssh_git_username` test to pass
`-o UserKnownHostsFile=/dev/null` to OpenSSH, in order to prevent it
from trying to write into the known-hosts in user's home directory.
To add insult to the injury, OpenSSH ignores `HOME` and writes into
the actual home when it is explicitly overridden for the purpose
of testing.