Support transparently upgradeable Python bin installations (#13531)
For `uv python install --preview` and `uv python install <minor-version>
--preview`, this installs transparently upgradeable executables into
`bin`. On macOS/Linux, it uses a symlink pointing to a path containing a
minor version directory symlink. On Windows, it uses a trampoline with a
junction-containing path as its target.

When creating `bin` links, if the install is either a default install or
a minor version install, we check if we can derive a `DirectorySymlink`
from each installation. If we can, then
* on macOS/Linux, we use the `DirectorySymlink` to link the `bin` link
to an executable path containing the symlink directory.
* on Windows, we use the `DirectorySymlink` (which contains a junction)
to create our trampoline.

On Windows, this also updates a couple of other things:
* When doing comparisons to determine if we have a bin link, if we have
a trampoline, we fully resolve its target path using
`dunce::canonicalize`.
* In the trampoline itself, if it's a Python trampoline, we check if
we're in a virtual environment (looking for a `pyvenv.cfg` with a `home`
key). If we are not in a virtual environment and the `PYTHONHOME` env
var is not set, we set `PYTHONHOME` to the parent directory of the
executable in order to ensure that `sys.path` is set with the correct
values.

This supports both upgradeable `python{major}.{minor}` and freethreaded
`python{major}.{minor}t` installs. `uv python upgrade 3.13` will upgrade
the former while `uv python upgrade 3.13t` will upgrade the latter.

This depends on PR #13712.